fix(frontend): total_asset long→double, current_price int→float — 소수 파싱 오류 수정 (#122, PR #204)
Critical 1 수정(소수점 평단가 파싱)으로 avg_price가 소수가 되면서
현재가 미상 종목의 total_asset 기여분(avg_price×quantity)도 소수가 됨.
Unity JsonUtility는 소수 JSON 값을 long 필드에 파싱하지 못하므로
PortfolioData.total_asset을 long→double로 변경한다.

models.py의 current_price: Optional[float] 스키마를 따라
Holding.current_price도 int→float으로 맞춘다.

GetWeight 파라미터를 long→double로 통일하고, 내부 계산을
double로 승격해 정수 오버플로를 방지한다.

+++ frontend/Assets/Scripts/PortfolioData.cs
@@ -2,7 +2,10 @@
 public class Holding
 {
     public string name;
-    public int current_price;
+    // models.py: current_price Optional[float] — DB 스키마가 소수를 허용하므로
+    // float으로 받아 소수 부분 손실을 방지한다.
+    // (한국 주식 현재가는 실제로 항상 원 단위 정수이지만 스키마 타입을 따른다.)
+    public float current_price;
     // API는 float(가중평균 매수단가)를 반환한다. int로 받으면 소수 부분이 잘린다.
     public float avg_price;
     public float return_rate;
@@ -12,18 +15,22 @@ public class Holding
     // price_known=false이면 current_price·return_rate는 0이 아니라 미확인 값이다.
     public bool price_known;
 
-    public float GetWeight(long total_asset)
+    public float GetWeight(double total_asset)
     {
         if (!price_known || total_asset == 0)
             return 0f;
-        return (float)(current_price * quantity) / total_asset * 100f;
+        // 정수 오버플로 방지: current_price(float)를 double로 승격한 뒤 계산한다.
+        return (float)((double)current_price * quantity / total_asset * 100.0);
     }
 }
 
 [System.Serializable]
 public class PortfolioData
 {
-    public long total_asset;
+    // avg_price(가중평균 매수단가)는 소수다. 현재가 미상 종목의 total_asset 기여분은
+    // avg_price × quantity로 계산되므로 total_asset도 소수가 될 수 있다.
+    // long으로 선언하면 Unity JsonUtility가 소수 JSON 값을 파싱하지 못한다.
+    public double total_asset;
     // true이면 total_asset에 현재가 없는 종목의 매입가 추정분이 포함된다.
     public bool total_asset_is_estimate;
     public float total_return_rate;
